feat: implement property extraction and UI components for inspector widget details
